Encode signals as a two-element sequence
Compared to packing the signal name and parameters into a single array,
this makes consuming signals more ergonomic. If a consumer first decodes
the signal name, it can then use an automatically generated decoding
routine to decode a single item (the parameters array) into a useful,
structured form, rather than having to decode the remaining portion of
the inside of an array.